improve the keybinding issue with Evil normal state
After this commit, the issue is resolved if Emacs performs native
compilation on `sigh-mode.el` and Emacs is restarted so that it loads
the compiled version of Sigh Mode.

It is unclear to me why this problem occurs on first load.  It is as
though `evil-normalize-keymaps` did not run unless it is Sigh Mode's
native compiled version that is loaded.